Improve simple vaultfile example
